refactor(cli): move command logic into modules
