+++ src/commands/mod.rs
@@ -1,0 +1,5 @@
+pub mod config;
+pub mod doctor;
+pub mod index;
+pub mod query;
+pub mod stat;
